add resolution for keytar and update to latest Theia next
The newer keytar version, pulled by default, does not rebuild correctly
(node-gyp). Aside from adding a resolution to pin to the earlier version,
Another way that it can be made to work is by building with a node version
that's aligned with the version of Electron we use. To find-out the
version, use `npx electron` in the repo root and read the Electron windows
launched. (ATM it's node v12.14.1).

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>